Add Stagging base support and add week close price compare test

+++ src/Stagging/SystemReport.java
@@ -30,7 +30,7 @@ public class SystemReport {
     public float blackWeightEarning;
     public float openWeightEarning;
     public float closeWeightEarning;
-    public float meanInvestCount;
+    public int meanInvestCount;
     public float meanDailyRate;
     public float meanPositionDays;
     public float maxInvest;
